send usecases to the backend branch

+++ api/src/modules/room/usecases/DeleteRoomUseCase.ts
@@ -8,8 +8,16 @@ export class DeleteRoomUseCase{
     async execute(roomId: number){
         const room = await this.roomRepository.findById(roomId);
         
+        if (!room) {
+            throw new Error("Sala não encontrada.");
+        }
+        
+        if (room.is_active) {
+            throw new Error("Apenas salas inativas podem ser deletadas.");
+        }        
+        
         // TODO: Need User Type Verification before delete!
-        // TODO: Must be inactive!
+        // TODO: Verify if the room has reservations before deleting.
         this.roomRepository.delete(roomId)
     }
 }
